+++ packages/xrpl/src/models/ledger/Amendments.ts
@@ -6,7 +6,7 @@ import BaseLedgerEntry from './BaseLedgerEntry'
 export const AMENDMENTS_ID =
   '7DB0788C020F02780A673DC74757F23823FA3014C1866E72CC4CD8B226CD6EF4'
 
-interface Majority {
+export interface Majority {
   Majority: {
     /** The Amendment ID of the pending amendment. */
     Amendment: string