+++ packages/xrpl/src/models/ledger/NFTokenPage.ts
@@ -1,21 +1,19 @@
 import BaseLedgerEntry from './BaseLedgerEntry'
 
 export interface NFToken {
-  Flags: number
-  Issuer: string
-  NFTokenID: string
-  NFTokenTaxon: number
-  URI?: string
-}
-
-export interface NFTokenWrapper {
-  NFToken: NFToken
+  NFToken: {
+    Flags: number
+    Issuer: string
+    NFTokenID: string
+    NFTokenTaxon: number
+    URI?: string
+  }
 }
 
 export interface NFTokenPage extends BaseLedgerEntry {
   LedgerEntryType: 'NFTokenPage'
   NextPageMin?: string
-  NFTokens: NFTokenWrapper[]
+  NFTokens: NFToken[]
   PreviousPageMin?: string
   PreviousTxnID?: string
   PreviousTxnLgrSeq?: number