+++ packages/xrpl/src/models/ledger/NegativeUNL.ts
@@ -6,11 +6,6 @@ import BaseLedgerEntry from './BaseLedgerEntry'
 export const NEGATIVE_UNL_ID =
   '2E8A59AA9D3B5B186B0B9E0F62E6C02587CA74A4D778938E957B6357D364B244'
 
-interface DisabledValidator {
-  FirstLedgerSequence: number
-  PublicKey: string
-}
-
 /**
  * The NegativeUNL object type contains the current status of the Negative UNL,
  * a list of trusted validators currently believed to be offline.
@@ -22,7 +17,10 @@ export default interface NegativeUNL extends BaseLedgerEntry {
   /**
    * A list of trusted validators that are currently disabled.
    */
-  DisabledValidators?: DisabledValidator[]
+  DisabledValidators?: Array<{
+    FirstLedgerSequence: number
+    PublicKey: string
+  }>
   /**
    * The public key of a trusted validator that is scheduled to be disabled in
    * the next flag ledger.